Sunday Matins Complete
The Template for Sunday Matins is Complete.
The first hour has been separated from teh Matins template and set in a template of its own.
The first hour template also contains the APPENDICES i.e. The resurrectional exaposteilaria, the 11 Eothina and the Polyeleos.

NB. The idea is to also separate the appendices to be one template for ease of access and re-use.